product-services/product_shipping_service/requirements.txt
Flask==1.1.2
jsonify==0.5
mysql-connector-python==8.0.21
requests==2.24.0
bson==0.5.10
waitress==1.4.4 
